requirements.txt
pyautogui
websocket-client
